+++ MemoryTrackerObj/Model/Constants.h
@@ -15,14 +15,14 @@ FOUNDATION_EXPORT NSString *const MTLevelIndex;
 FOUNDATION_EXPORT NSString *const MTSoundState;
 FOUNDATION_EXPORT NSString *const MTMelodyType;
 
-FOUNDATION_EXPORT int const MTDefaultTimeLimit;
-FOUNDATION_EXPORT int const MTScoreIncrementVal;
-FOUNDATION_EXPORT int const MTScoreDecrementVal;
-FOUNDATION_EXPORT int const MTScoreMultiplier;
+FOUNDATION_EXPORT NSInteger const MTDefaultTimeLimit;
+FOUNDATION_EXPORT NSInteger const MTScoreIncrementVal;
+FOUNDATION_EXPORT NSInteger const MTScoreDecrementVal;
+FOUNDATION_EXPORT NSInteger const MTScoreMultiplier;
 
-FOUNDATION_EXPORT int const MTRawsCountInDesk;
-FOUNDATION_EXPORT int const MTColumnsCountInDesk;
-FOUNDATION_EXPORT int const MTCardsCount;
+FOUNDATION_EXPORT NSInteger const MTRawsCountInDesk;
+FOUNDATION_EXPORT NSInteger const MTColumnsCountInDesk;
+FOUNDATION_EXPORT NSInteger const MTCardsCount;
 
 
 #endif /* Constants_h */
